Implemented Calls for Friendships, All Users and Stats in the UI

+++ app/src/main/java/com/example/quizit_android_app/model/DataClasses.kt
@@ -40,7 +40,7 @@ data class UserStatsResponse (
 data class Stats (
     @SerializedName("avgPoints" ) var avgPoints : Double? = null,
     @SerializedName("ranking"   ) var ranking   : Int?    = null,
-    @SerializedName("winRate"   ) var winRate   : Int?    = null
+    @SerializedName("winRate"   ) var winRate   : Double?    = null
 )
